AML-3882: migrate WhenIGetAccountUsersWithUnkownIds; except for the ExpectedControllerType assertion.
