+++ src/SFA.DAS.EmployerAccounts.Api.IntegrationTests/SFA.DAS.EmployerAccounts.Api.IntegrationTests.csproj
@@ -301,6 +301,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="GivenEmployerAccountsApi\EmployerAccountControllerTests\WhenGetAccountUsersWithUnknownIds.cs" />
     <Compile Include="GivenEmployerAccountsApi\GivenEmployerAccountsApi.cs" />
     <Compile Include="GivenEmployerAccountsApi\LegalEntitiesControllerTests\WhenIGetASingleLegalEntityWithKnownIds.cs" />
     <Compile Include="GivenEmployerAccountsApi\LegalEntitiesControllerTests\WhenIGetLegalEntitiesWithUnknownIds.cs" />
